+++ apps/fluent-tester/src/FluentTester/TestComponents/Menu/MenuTest.tsx
@@ -26,9 +26,9 @@ const MenuDefault: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItem content="A plain MenuItem" />
-            <MenuItem disabled content="A second disabled plain MenuItem" />
-            <MenuItem content="A third plain MenuItem" />
+            <MenuItem>A plain MenuItem</MenuItem>
+            <MenuItem disabled>A second disabled plain MenuItem</MenuItem>
+            <MenuItem>A third plain MenuItem</MenuItem>
           </MenuList>
         </MenuPopover>
       </Menu>
@@ -45,9 +45,9 @@ const MenuCheckmarks: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItemCheckbox name="itemOne" content="A MenuItem with checkmark" />
+            <MenuItemCheckbox name="itemOne">A MenuItem with checkmark</MenuItemCheckbox>
             <MenuDivider />
-            <MenuItemCheckbox name="itemTwo" content="Another MenuItem with checkmark" />
+            <MenuItemCheckbox name="itemTwo">Another MenuItem with checkmark</MenuItemCheckbox>
           </MenuList>
         </MenuPopover>
       </Menu>
@@ -57,10 +57,12 @@ const MenuCheckmarks: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItem content="A plain MenuItem" />
-            <MenuItemCheckbox name="itemTwo" content="A MenuItem with checkmark" />
-            <MenuItemCheckbox disabled name="itemThree" content="A disabled MenuItem with checkmark" />
-            <MenuItemCheckbox name="itemFour" content="A MenuItem with checkmark" />
+            <MenuItem>A plain MenuItem</MenuItem>
+            <MenuItemCheckbox name="itemTwo">A MenuItem with checkmark</MenuItemCheckbox>
+            <MenuItemCheckbox disabled name="itemThree">
+              A disabled MenuItem with checkmark
+            </MenuItemCheckbox>
+            <MenuItemCheckbox name="itemFour">A MenuItem with checkmark</MenuItemCheckbox>
           </MenuList>
         </MenuPopover>
       </Menu>
@@ -70,9 +72,9 @@ const MenuCheckmarks: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItem content="A plain MenuItem" />
-            <MenuItemCheckbox name="itemTwo" content="A MenuItem with checkmark" />
-            <MenuItemCheckbox name="itemThree" content="A MenuItem with checkmark" />
+            <MenuItem>A plain MenuItem</MenuItem>
+            <MenuItemCheckbox name="itemTwo">A MenuItem with checkmark</MenuItemCheckbox>
+            <MenuItemCheckbox name="itemThree">A MenuItem with checkmark</MenuItemCheckbox>
           </MenuList>
         </MenuPopover>
       </Menu>
@@ -98,10 +100,10 @@ const MenuRadioItem: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItemRadio name="itemOne" content="A MenuItem with checkmark and radio selection" />
-            <MenuItemRadio name="itemTwo" content="Another MenuItem with checkmark and radio selection" />
-            <MenuItemRadio name="itemThree" content="A third MenuItem with checkmark and radio selection" />
-            <MenuItemCheckbox name="itemFour" content="A MenuItem with checkmark and toggle selection" />
+            <MenuItemRadio name="itemOne">A MenuItem with checkmark and radio selection</MenuItemRadio>
+            <MenuItemRadio name="itemTwo">Another MenuItem with checkmark and radio selection</MenuItemRadio>
+            <MenuItemRadio name="itemThree">A third MenuItem with checkmark and radio selection</MenuItemRadio>
+            <MenuItemCheckbox name="itemFour">A MenuItem with checkmark and toggle selection</MenuItemCheckbox>
           </MenuList>
         </MenuPopover>
       </Menu>
@@ -113,12 +115,12 @@ const Submenu: React.FunctionComponent = () => {
   return (
     <Menu>
       <MenuTrigger>
-        <MenuItem content="A second MenuItem" />
+        <MenuItem>A second MenuItem</MenuItem>
       </MenuTrigger>
       <MenuPopover>
         <MenuList>
-          <MenuItem content="A nested MenuItem" />
-          <MenuItem content="A second nested MenuItem" />
+          <MenuItemCheckbox name={'a'}>A nested MenuItemCheckbox</MenuItemCheckbox>
+          <MenuItem>A nested MenuItem</MenuItem>
           <Submenu />
         </MenuList>
       </MenuPopover>
@@ -135,7 +137,7 @@ const MenuSubMenu: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItem content="A MenuItem" />
+            <MenuItem>A MenuItem</MenuItem>
             <Submenu />
           </MenuList>
         </MenuPopover>
@@ -153,7 +155,7 @@ const MenuOpenOnHover: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItem content="A MenuItem" />
+            <MenuItem>A MenuItem</MenuItem>
             <Submenu />
           </MenuList>
         </MenuPopover>
@@ -173,10 +175,61 @@ const MenuControlledOpen: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <MenuItem content="A MenuItem" />
+            <MenuItem>A MenuItem</MenuItem>
             <Submenu />
           </MenuList>
         </MenuPopover>
+      </Menu>
+    </Stack>
+  );
+};
+
+const CustomMenuPopover = MenuPopover.customize({ borderWidth: 4, borderColor: 'blue' });
+const CustomMenuList = MenuList.customize({ gap: 5, padding: 6, backgroundColor: 'pink' });
+const CustomMenuItem = MenuItem.customize({
+  gap: 5,
+  padding: 8,
+  backgroundColor: 'red',
+  color: 'white',
+  borderRadius: 2,
+  checkmarkSize: 10,
+  submenuIndicatorPadding: 4,
+  submenuIndicatorSize: 10,
+  fontSize: 14,
+});
+const CustomMenuCheckmark = MenuItemCheckbox.customize({
+  gap: 5,
+  padding: 8,
+  backgroundColor: 'red',
+  color: 'white',
+  borderRadius: 2,
+  checkmarkSize: 10,
+  fontSize: 14,
+  checkmarkColor: 'blue',
+});
+const CustomMenuDivider = MenuDivider.customize({
+  height: 8,
+  margin: 4,
+  marginVertical: 6,
+  width: 100,
+  backgroundColor: 'purple',
+});
+
+const MenuCustomized: React.FunctionComponent = () => {
+  return (
+    <Stack style={stackStyle}>
+      <Menu>
+        <MenuTrigger>
+          <Button>Test</Button>
+        </MenuTrigger>
+        <CustomMenuPopover>
+          <CustomMenuList>
+            <CustomMenuItem>A MenuItem</CustomMenuItem>
+            <CustomMenuItem>A MenuItem</CustomMenuItem>
+            <CustomMenuDivider />
+            <CustomMenuCheckmark name={'custom'}>A MenuItem</CustomMenuCheckmark>
+          </CustomMenuList>
+        </CustomMenuPopover>
       </Menu>
     </Stack>
   );
@@ -207,6 +260,10 @@ const menuSections: TestSection[] = [
   {
     name: 'Menu Submenu',
     component: MenuSubMenu,
+  },
+  {
+    name: 'Menu Customized',
+    component: MenuCustomized,
   },
   {
     name: 'Menu E2E',