+++ apps/fluent-tester/src/FluentTester/testPages.windows.ts
@@ -1,4 +1,4 @@
-import { JSAvatarTest, HOMEPAGE_AVATAR_BUTTON } from './TestComponents/Avatar';
+import { AvatarTest, HOMEPAGE_AVATAR_BUTTON } from './TestComponents/Avatar';
 import { ButtonTest, HOMEPAGE_BUTTON_BUTTON } from './TestComponents/Button';
 import { BadgeTest, HOMEPAGE_BADGE } from './TestComponents/Badge';
 import { CalloutTest, HOMEPAGE_CALLOUT_BUTTON } from './TestComponents/Callout';
@@ -25,7 +25,7 @@ export const tests = [
   },
   {
     name: 'Avatar Test',
-    component: JSAvatarTest,
+    component: AvatarTest,
     testPage: HOMEPAGE_AVATAR_BUTTON,
   },
   {