+++ packages/experimental/Badge/src/PresenceBadge/PresenceBadge.styling.ts
@@ -1,41 +1,58 @@
 import { PresenceBadgeTokens, PresenceBadgeSlotProps, PresenceBadgeProps, presenceBadgeName } from './PresenceBadge.types';
 import { UseStylingOptions, buildProps, Theme } from '@fluentui-react-native/framework';
-import { borderStyles, layoutStyles } from '@fluentui-react-native/tokens';
+import { borderStyles } from '@fluentui-react-native/tokens';
 import { defaultBadgeTokens } from '../BadgeTokens';
 import { defaultPresenceBadgeTokens } from './PresenceBadgeTokens';
 import { coreBadgeStates, getBadgePosition } from '../Badge.styling';
 
+export const badgeStates: (keyof PresenceBadgeTokens)[] = [
+  ...coreBadgeStates,
+  'doNotDisturb',
+  'busy',
+  'unknown',
+  'blocked',
+  'outOfOffice',
+  'away',
+  'available',
+  'offline',
+  'awayOutOfOffice',
+];
+
 export const stylingSettings: UseStylingOptions<PresenceBadgeProps, PresenceBadgeSlotProps, PresenceBadgeTokens> = {
   tokens: [defaultBadgeTokens, presenceBadgeName, defaultPresenceBadgeTokens],
-  states: coreBadgeStates,
+  states: badgeStates,
   slotProps: {
     root: buildProps(
-      (tokens: PresenceBadgeTokens, theme: Theme) => ({
-        style: {
-          ...getBadgePosition(tokens),
-          width: tokens.width,
-          height: tokens.height,
-          display: 'flex',
-          alignItems: 'center',
-          flexDirection: 'row',
-          alignSelf: 'flex-start',
-          justifyContent: 'center',
-          position: 'absolute',
-          backgroundColor: tokens.backgroundColor,
-          ...borderStyles.from(tokens, theme),
-          ...layoutStyles.from(tokens, theme),
-        },
-      }),
-      ['backgroundColor', 'width', 'height', 'bottom', 'right', 'top', 'left', ...borderStyles.keys, ...layoutStyles.keys],
+      (tokens: PresenceBadgeTokens, theme: Theme) => {
+        const { width, height, borderWidth } = tokens;
+        const borderGap = borderWidth * 2;
+        return {
+          style: {
+            ...getBadgePosition(tokens),
+            width: width + borderGap,
+            height: height + borderGap,
+            alignItems: 'center',
+            flexDirection: 'row',
+            alignSelf: 'flex-start',
+            justifyContent: 'center',
+            position: 'absolute',
+            backgroundColor: tokens.backgroundColor,
+            ...borderStyles.from(tokens, theme),
+            paddingHorizontal: tokens.paddingHorizontal,
+          },
+        };
+      },
+      ['backgroundColor', 'width', 'height', 'bottom', 'right', 'top', 'left', 'paddingHorizontal', ...borderStyles.keys],
     ),
-    svgXml: buildProps(
+    svg: buildProps(
       (tokens: PresenceBadgeTokens) => ({
         style: {
           width: tokens.width,
           height: tokens.height,
+          color: tokens.iconColor,
         },
       }),
-      ['width', 'height'],
+      ['width', 'height', 'iconColor'],
     ),
   },
 };