+++ packages/experimental/Menu/src/consts.ts
@@ -1,8 +1,0 @@
-import { Platform } from 'react-native';
-
-export const delayHover = Platform.select({
-  macos: 100,
-  default: 500, // win32
-});
-
-export const isCloseOnHoverOutEnabled = Platform.OS === ('win32' as any);