+++ packages/utils/adapters/src/adapters.windows.ts
@@ -98,6 +98,8 @@ const _textMask: IFilterMask<ITextProps> = {
   onLongPress: true,
   onMagicTap: true,
   onPress: true,
+  onPressIn: true,
+  onPressOut: true,
   style: true,
   nativeID: true,
   maxFontSizeMultiplier: true,