+++ packages/deprecated/foundation-tokens/CHANGELOG.md
@@ -1,8 +1,28 @@
 # Change Log - @uifabricshared/foundation-tokens
 
-This log was last generated on Thu, 31 Mar 2022 07:27:48 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Jul 2022 18:09:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.0
+
+Thu, 14 Jul 2022 18:09:50 GMT
+
+### Minor changes
+
+- Update to React Native 0.66 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/merge-props to v0.5.0
+- Bump @fluentui-react-native/tokens to v0.16.0
+- Bump @uifabricshared/foundation-settings to v0.12.0
+
+## 0.11.4
+
+Fri, 08 Jul 2022 21:23:36 GMT
+
+### Patches
+
+- Move used tokens types to tokens package (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/tokens to v0.15.1
 
 ## 0.11.3