+++ packages/theming/android-theme/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui-react-native/android-theme
 
-This log was last generated on Wed, 25 May 2022 18:43:07 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Jul 2022 18:09:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.0
+
+Thu, 14 Jul 2022 18:09:50 GMT
+
+### Minor changes
+
+- Update to React Native 0.66 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/theme-types to v0.18.0
+- Bump @fluentui-react-native/theme to v0.7.0
 
 ## 0.10.0